알람 확인 기능을 추가했습니다.!
