+++ gaia/webapp/resources/assets/css/alarm.css
@@ -7,8 +7,6 @@
 }
 #alarmList{
 	padding : 0;
-}
-.newnotitag{
-    font-size: 1rem;
-    color: #333;
+	max-height : 400px;
+	overflow-y : auto;
 }
