+++ gaia/webapp/resources/assets/css/style.css
@@ -11909,20 +11909,20 @@ sub {
   display: block;
 }
 
-.dropdown-notfication li:nth-child(1) .avatar-icon {
-  background-color: #f6514c;
-}
+/* .dropdown-notfication li:nth-child(1) .avatar-icon { */
+/*   background-color: #f6514c; */
+/* } */
 
-.dropdown-notfication li:nth-child(2) .avatar-icon {
-  background-color: #449036;
-}
+/* .dropdown-notfication li:nth-child(2) .avatar-icon { */
+/*   background-color: #449036; */
+/* } */
 
-.dropdown-notfication li:nth-child(3) .avatar-icon {
-  background-color: #ce2b23;
-}
+/* .dropdown-notfication li:nth-child(3) .avatar-icon { */
+/*   background-color: #ce2b23; */
+/* } */
 
-.dropdown-notfication li:nth-child(4) .avatar-icon {
-  background-color: #3375e1;
+/* .dropdown-notfication li:nth-child(4) .avatar-icon { */
+/*   background-color: #3375e1; */
 }
 
 .dropdown-notfication .avatar-icon {
@@ -11932,6 +11932,7 @@ sub {
   justify-content: center;
   align-items: center;
   color: #fff;
+  background-color: #BBBBBB;
 }
 
 .dropdown-mega-menu {
